# apps/api/src/focusdive_api/api/v1/integrations/slack/schemas.py
from typing import Literal

from pydantic import BaseModel


class SlackStatusOut(BaseModel):
    is_connected: bool


class SlackTestIn(BaseModel):
    action: Literal["start", "stop"]
    dnd_text: str | None = None
    dnd_emoji: str | None = None


class SlackTestOut(BaseModel):
    success: bool
    message: str | None = None