# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FocusDive API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/integrations/slack/status": {
      "get": {
        "tags": [
          "integrations:slack"
        ],
        "summary": "Slack Status",
        "operationId": "slack_status_v1_integrations_slack_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlackStatusOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/integrations/slack/test": {
      "post": {
        "tags": [
          "integrations:slack"
        ],
        "summary": "Slack Test",
        "operationId": "slack_test_v1_integrations_slack_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SlackTestIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlackTestOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SlackStatusOut": {
        "properties": {
          "is_connected": {
            "type": "boolean",
            "title": "Is Connected"
          }
        },
        "type": "object",
        "required": [
          "is_connected"
        ],
        "title": "SlackStatusOut"
      },
      "SlackTestIn": {
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "start",
              "stop"
            ],
            "title": "Action"
          },
          "dnd_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dnd Text"
          },
          "dnd_emoji": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dnd Emoji"
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "SlackTestIn"
      },
      "SlackTestOut": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "SlackTestOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}